Move CT related classes into new package
