+++ base/ca/src/org/dogtagpki/ct/CTRequest.java
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 //
 
-package com.netscape.ca;
+package org.dogtagpki.ct;
 
 import java.util.ArrayList;
 import java.util.List;
